ログイン機能とタスク作成

+++ app/Http/Controllers/TasksController.php
@@ -16,13 +16,23 @@ class TasksController extends Controller
     //getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        //タスク一覧を取得
-        $tasks = Task::all();
+        
+        
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
         
         //タスク一覧表示
-        return view('tasks.index', [
-            'tasks' => $tasks, 
-            ]);
+        return view('tasks.index', $data);
     }
 
     /**
@@ -57,11 +67,11 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
-        //タスクを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        //認証済みユーザのタスク作成としてデータ作成、保存
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content'=> $request->content,
+        ]);
         
         //トップページへリダイレクト
         return redirect('/');
@@ -120,10 +130,15 @@ class TasksController extends Controller
         
         //idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        //タスクを更新
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        
+        //認証済みユーザがそのタスク所有者である場合、そのタスクを削除
+        if (\Auth::id() === $task->user_id) {
+            $request->user()->tasks()->update([
+                'status' => $request->status,
+                'content'=> $request->content,
+            ]);
+            
+        }
 
         //トップページへリダイレクト
         return redirect('/');
@@ -140,8 +155,11 @@ class TasksController extends Controller
     {
         //idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        //タスクを削除
-        $task->delete();
+        
+        //認証済みユーザがそのタスク所有者である場合、そのタスクを削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
